app/library/utils.py
import torch

def generate(prompt, max_seq_len, temperature, model, tokenizer, vocab, device):
    """
    Generates text based on the given prompt using the LSTM language model.

    Args:
        prompt (str): Initial text prompt.
        max_seq_len (int): Maximum length of the generated sequence.
        temperature (float): Controls randomness in text generation.
        model (torch.nn.Module): Trained LSTM model.
        tokenizer (callable): Tokenizer function to split the input.
        vocab (torchtext.vocab.Vocab): Vocabulary for token-to-index conversion.
        device (torch.device): Device to run the model on.

    Returns:
        list: Generated tokens.
    """
    model.eval()
    tokens = tokenizer(prompt)
    input_ids = [vocab[token] if token in vocab else vocab['<unk>'] for token in tokens]
    input_tensor = torch.tensor(input_ids).unsqueeze(0).to(device)

    output_tokens = tokens.copy()  # Start with the prompt tokens
    with torch.no_grad():
        for _ in range(max_seq_len):
            logits, _ = model(input_tensor)
            next_token_logits = logits[:, -1, :] / temperature
            probabilities = torch.softmax(next_token_logits, dim=-1)
            next_token_id = torch.multinomial(probabilities, num_samples=1).item()

            if next_token_id == vocab['<eos>']:
                break

            next_token = vocab.lookup_token(next_token_id)
            output_tokens.append(next_token)

            input_tensor = torch.cat([input_tensor, torch.tensor([[next_token_id]], device=device)], dim=1)

    return output_tokens
